Moved data linking classes to own files
