+++ tests/test_scoring.py
@@ -4,7 +4,7 @@ from nplinker.genomics import BGC
 from nplinker.genomics import GCF
 from nplinker.metabolomics import Spectrum
 from nplinker.scoring.linking.data_linking import DataLinks
-from nplinker.scoring.linking.data_linking import LinkFinder
+from nplinker.scoring.linking.link_finder import LinkFinder
 from nplinker.scoring.misc import hg_scoring
 from nplinker.scoring.misc import metcalf_scoring
 from nplinker.strains import Strain
